feat: initialize map in app component
